# services/api/src/main/resources/META-INF/openapi-v1.yaml
openapi: 3.0.3
info:
  title: JavaBlog API
  description: REST API for javablog.com blog posts and comments
  version: 1.0.0
servers:
  - url: /v1
    description: API version 1
...
paths:
...
  /posts/{postId}/translations:
    get:
      tags:
        - Translations
      operationId: listTranslations
      summary: List translation state for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of translated posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslatedPost'
        '404':
          description: Post not found
    post:
      tags:
        - Translations
      operationId: requestTranslation
      summary: Request translation of a post into one or more languages
      security:
        - oidc: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRequestBody'
      responses:
        '204':
          description: Translation request accepted
        '400':
          description: Invalid request - requested languages contain the original post language
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Post not found
...
  /articles:
    post:
      tags:
        - Articles
      operationId: createArticle
      summary: Create a new how-to article
      security:
        - oidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleRequest'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '401':
          description: Unauthorized - valid JWT required
    get:
      tags:
        - Articles
      parameters:
        - name: language
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/LanguageCode'
      operationId: listArticles
      summary: List all how-to articles sorted by newest first
      responses:
        '200':
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleResponse'
  /articles/by-slug/{slug}:
    get:
      tags:
        - Articles
      operationId: getArticleBySlug
      summary: Get a how-to article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          description: Article not found
  /articles/{articleId}:
    get:
      tags:
        - Articles
      operationId: getArticle
      summary: Get a how-to article by ID
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Article found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          description: Article not found
    put:
      tags:
        - Articles
      operationId: updateArticle
      summary: Update a how-to article
      security:
        - oidc: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleRequest'
      responses:
        '200':
          description: Article updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Article not found
  /articles/{articleId}/translations:
    get:
      tags:
        - ArticleTranslations
      operationId: listArticleTranslations
      summary: List translation state for an article
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of translated articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslatedArticle'
        '404':
          description: Article not found
    post:
      tags:
        - ArticleTranslations
      operationId: requestArticleTranslation
      summary: Request translation of an article into one or more languages
      security:
        - oidc: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleTranslationRequestBody'
      responses:
        '204':
          description: Translation request accepted
        '400':
          description: Invalid request
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Article not found
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://keycloak.javablog.com/realms/javablog/.well-known/openid-configuration
      description: Keycloak OIDC authentication
  schemas:
    LanguageCode:
      type: string
      description: ISO 639-1 two-letter code (lowercase) for supported languages.
      enum:
        - bg   # Български
        - hr   # Hrvatski
        - cs   # Čeština
        - da   # Dansk
        - nl   # Nederlands
        - en   # English
        - et   # Eesti
        - fi   # Suomi
        - fr   # Français
        - de   # Deutsch
        - el   # Ελληνικά
        - hu   # Magyar
        - ga   # Gaeilge
        - it   # Italiano
        - lv   # Latviešu
        - lt   # Lietuvių
        - mt   # Malti
        - pl   # Polski
        - pt   # Português
        - ro   # Română
        - sk   # Slovenčina
        - sl   # Slovenščina
        - es   # Español
        - sv   # Svenska
        - ru   # Русский
        - uk   # Українська
        - be   # Беларуская
        - sq   # Shqip
        - "no" # Norsk
        - is   # Íslenska
        - ca   # Català
        - eu   # Euskara
        - gd   # Gàidhlig
        - cy   # Cymraeg
        - zh   # 简体中文
        - tw   # 臺灣華語
        - hi   # हिन्दी
        - ar   # العربية
        - id   # Bahasa Indonesia
        - bn   # বাংলা
        - ja   # 日本語
        - ko   # 한국어
        - tr   # Türkçe
        - he   # עברית
    TranslatedPost:
      type: object
      required:
        - orderedDatetime
        - language
      properties:
        orderedDatetime:
          type: string
          format: date-time
          description: When the translation was requested
        language:
          $ref: '#/components/schemas/LanguageCode'
        completedDatetime:
          type: string
          format: date-time
          description: When the translation was completed
        title:
          type: string
          description: Translated post title
        slug:
          type: string
          description: Translated post slug
    TranslationRequestBody:
      type: object
      required:
        - languages
      properties:
        languages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LanguageCode'
          description: Target languages for translation
...
    CreateArticleRequest:
      type: object
      required:
        - slug
        - title
        - summary
        - content
        - language
      properties:
        slug:
          type: string
          description: URL-friendly identifier for the article
        title:
          type: string
          description: Article title
        summary:
          type: string
          maxLength: 500
          description: Short summary for listings
        content:
          type: string
          description: Article content
        language:
          $ref: '#/components/schemas/LanguageCode'
    UpdateArticleRequest:
      type: object
      required:
        - slug
        - title
        - summary
        - content
        - language
      properties:
        slug:
          type: string
          description: URL-friendly identifier for the article
        title:
          type: string
          description: Article title
        summary:
          type: string
          maxLength: 500
          description: Short summary for listings
        content:
          type: string
          description: Article content
        language:
          $ref: '#/components/schemas/LanguageCode'
    ArticleResponse:
      type: object
      required:
        - id
        - slug
        - title
        - summary
        - content
        - language
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique article identifier
        slug:
          type: string
          description: URL-friendly identifier
        title:
          type: string
          description: Article title
        summary:
          type: string
          description: Short summary for listings
        content:
          type: string
          description: Article content
        language:
          $ref: '#/components/schemas/LanguageCode'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
    TranslatedArticle:
      type: object
      required:
        - orderedDatetime
        - language
      properties:
        orderedDatetime:
          type: string
          format: date-time
          description: When the translation was requested
        language:
          $ref: '#/components/schemas/LanguageCode'
        completedDatetime:
          type: string
          format: date-time
          description: When the translation was completed
        title:
          type: string
          description: Translated article title
        slug:
          type: string
          description: Translated article slug
    ArticleTranslationRequestBody:
      type: object
      required:
        - languages
      properties:
        languages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LanguageCode'
          description: Target languages for translation